move jnit setup into buildsrc
