+++ gradle.properties
@@ -18,7 +18,7 @@ kotlin.code.style=official
 org.gradle.daemon=true
 org.gradle.configureondemand=true
 org.gradle.jvmargs=-Xmx4g -XX:MaxPermSize=2048m -Dfile.encoding=UTF-8
-kotlin.native.cacheKind=none
+# kotlin.native.cacheKind=none
 kotlin.native.ignoreDisabledTargets=true
 android.useAndroidX=true
 android.enableJetifier=true
